get_images.py obtains a random image from imgur
This is a very rough prototype, mostly designed as a proof of concept.
Currently, a file named client_id.txt must be included, containing
nothing but the client id with an ending newline. This file was not
included currently, because I'm not certain if client ids are supposed
to be publicly available. I will probably turn this into a passed in
parameter later on anyway.